Comment form and blog front end

+++ app/Http/Controllers/Blogs/BlogController.php
@@ -5,12 +5,17 @@ use App\Http\Controllers\Controller;
 
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Auth;
+use Illuminate\Foundation\Auth\Access\AuthorizesRequests;
 
 use App\Models\Blog;
+
+use League\CommonMark\CommonMarkConverter;
 
 
 class BlogController extends Controller
 {
+    use AuthorizesRequests;
+
     public function index(Request $request)
     {
         $blogs = Blog::query();
@@ -35,7 +40,6 @@ class BlogController extends Controller
         ]);
     }
 
-
     public function create()
     {
         return view('blogs.create');
@@ -59,7 +63,20 @@ class BlogController extends Controller
 
     public function show(Blog $blog)
     {
-        $blog->load('owner', 'comments.replies');
+        // Load comments, top-level only, with their replies and commenters
+        $blog->load(['comments' => function ($query) {
+            $query->whereNull('parent_id')->latest()
+                ->with(['commenter', 'replies.commenter']);
+        }]);
+    
+        // Convert Markdown to HTML
+        $converter = new CommonMarkConverter([
+            'html_input' => 'escape',
+            'allow_unsafe_links' => false,
+        ]);
+    
+        $blog->parsed_content = $converter->convertToHtml($blog->content);
+    
         return view('blogs.show', compact('blog'));
     }
 
@@ -85,7 +102,7 @@ class BlogController extends Controller
 
     public function destroy(Blog $blog)
     {
-        //$this->authorize('delete', $blog);
+        $this->authorize('delete', $blog);
         $blog->delete();
         return redirect()->route('blogs.index')->with('success', 'Blog deleted!');
     }
